requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
pydantic-settings>=2.2
httpx>=0.27
numpy>=1.26
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pypdf>=4.2.0
python-docx>=1.1.2
tqdm>=4.66
